relocated Comment id into id-types.ts
